refactor: organize code and document why
