+++ internal/acceptance/workflows/scenario/shared_state.go
