+++ internal/acceptance/workflows/scenario/step_funcs_ops_manager.go
@@ -2,6 +2,8 @@ package scenario
 
 import (
 	"context"
+	"fmt"
+	"github.com/pivotal-cf/kiln/pkg/cargo"
 	"os/exec"
 )
 
@@ -41,4 +43,33 @@ func iUploadConfigureAndApplyTheTile(ctx context.Context) (context.Context, erro
 	}
 
 	return ctx, nil
+}
+
+func theStemcellVersionInTheLockMatchesTheUsedForTheTile(ctx context.Context) error {
+	lockPath, err := kilnfileLockPath(ctx)
+	if err != nil {
+		return err
+	}
+
+	var stemcellAssociations struct {
+		StemcellLibrary []struct {
+			Version string `yaml:"version"`
+		} `yaml:"stemcell_library"`
+	}
+	err = runAndParseStdoutAsYAML(ctx,
+		exec.Command("om", "--skip-ssl-validation",
+			"curl", "--path", "/api/v0/stemcell_associations",
+		),
+		&stemcellAssociations,
+	)
+	if len(stemcellAssociations.StemcellLibrary) == 0 {
+		return fmt.Errorf("no stemcells found on ops manager")
+	}
+	var kl cargo.KilnfileLock
+	err = loadFileAsYAML(lockPath, &kl)
+	if err != nil {
+		return err
+	}
+	kl.Stemcell.Version = stemcellAssociations.StemcellLibrary[0].Version
+	return saveAsYAML(lockPath, kl)
 }
